refactor: separate pure image-tag/policy logic from provenance's network I/O
