+++ core/lib/provenance/verify-image.ts
@@ -15,24 +15,14 @@ import {
   fetchBundle,
   readGhcrBasicAuth,
 } from "./oci-registry.ts";
-import { derUtf8, verifyBundle, type VerifyBundleOptions, type DsseBundle } from "./sigstore.ts";
+import { verifyBundle } from "./sigstore.ts";
+import type { DsseBundle } from "./signed-digest.ts";
+import { imageTagFromRef } from "./image-tag.ts";
+import { buildVerifyOptions, type VerifyImageIdentity } from "./verify-policy.ts";
 import { ProvenanceError, VerifyImageError } from "./errors.ts";
 import { errorMessage } from "../general/error-message.ts";
 
 const REGISTRY = "ghcr.io";
-const EXPECTED_ISSUER = "https://token.actions.githubusercontent.com";
-const RELEASE_WORKFLOW = ".github/workflows/docker-publish.yml";
-
-// Fulcio OID: Source Repository Digest — the commit SHA of the build source.
-// Value encoding: DER UTF8String ([0x0C, len, ...utf8bytes]) inside OCTET STRING.
-const OID_SOURCE_REPO_DIGEST = "1.3.6.1.4.1.57264.1.13";
-
-const escapeRegex = (s: string): string => s.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-
-export interface VerifyImageIdentity {
-  actionRef: string;
-  actionRepo: string;
-}
 
 export interface VerifyImageDigestOptions extends VerifyImageIdentity {
   proxyEngine?: string;
@@ -42,80 +32,6 @@ export interface VerifyImageDigestOptions extends VerifyImageIdentity {
 export interface ResolvedImage {
   imageRef: string;
   pullPolicy: "always";
-}
-
-/**
- * Convert an action ref into the base Docker image tag, then append the
- * proxy engine suffix for non-default engines. The `transparent` engine
- * (default) publishes the plain version tag (e.g. `2.1.0`), matching the
- * pre-multi-engine tagging scheme; `explicit` (experimental) and `proxy`
- * (the buildkitd-less network-isolation proxy used by the run action)
- * each publish under their own suffix (e.g. `2.1.0-explicit`,
- * `2.1.0-proxy`). All three share the same Sigstore verification identity
- * (same workflow, same git ref) — only the published Docker tag differs, so
- * this does not affect buildVerifyOptions below.
- */
-export function imageTagFromRef(
-  actionRef: string | undefined,
-  proxyEngine: string = "transparent",
-): string {
-  if (!actionRef) return "";
-  let base;
-  if (/^[0-9a-f]{40}$/i.test(actionRef)) {
-    base = `sha-${actionRef.toLowerCase()}`;
-  } else if (actionRef.startsWith("v")) {
-    base = actionRef.slice(1);
-  } else {
-    base = actionRef;
-  }
-  if (proxyEngine === "explicit" || proxyEngine === "proxy") return `${base}-${proxyEngine}`;
-  return base;
-}
-
-/**
- * Build verify options encoding the expected certificate identity.
- *
- * The SAN URI pattern uses `(\.|$)` boundary anchors for version tags so that
- * e.g. @v2.1 matches v2.1.0 and v2.1.3 but NOT v2.10.0.
- *
- * For SHA pins, OID 1.13 (Source Repository Digest) pins the exact commit
- * while the SAN accepts any release tag.
- *
- * Returns null for unverifiable refs (branch names, local paths).
- */
-export function buildVerifyOptions({
-  actionRef,
-  actionRepo,
-}: VerifyImageIdentity): VerifyBundleOptions | null {
-  const sanPrefix = `^${escapeRegex(`https://github.com/${actionRepo}/${RELEASE_WORKFLOW}@refs/tags/`)}`;
-  const base = {
-    certificateIssuer: EXPECTED_ISSUER,
-    tlogThreshold: 1,
-    ctLogThreshold: 1,
-  };
-
-  // SHA pin: the SAN accepts any v*-prefixed release tag — the exact commit is
-  // pinned by OID 1.13 (Source Repository Digest), which enforces a strict byte
-  // match against the pinned SHA and cannot be satisfied by any other commit.
-  if (/^[0-9a-f]{40}$/i.test(actionRef)) {
-    return {
-      ...base,
-      certificateIdentityURI: `${sanPrefix}v`,
-      certificateOIDs: {
-        [OID_SOURCE_REPO_DIGEST]: derUtf8(actionRef.toLowerCase()),
-      },
-    };
-  }
-
-  // Version tag: SAN ref must match this version (boundary-safe via (\.|$)).
-  if (actionRef.startsWith("v")) {
-    return {
-      ...base,
-      certificateIdentityURI: `${sanPrefix}${escapeRegex(actionRef)}(\\.|$)`,
-    };
-  }
-
-  return null; // branch name, local ./setup, etc. — no verifiable release bundle
 }
 
 /**
